third_party/WebKit/LayoutTests/bluetooth/server/getPrimaryService/gen-delayed-discovery-service-not-found.html

<!-- Generated by //third_party/WebKit/LayoutTests/bluetooth/generate.py -->
<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../../resources/bluetooth/web-bluetooth-test.js"></script>
<script src="../../../resources/mojo-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  return setBluetoothFakeAdapter('DelayedServicesDiscoveryAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['heart_rate']}],
      optionalServices: ['battery_service']}))
    .then(device => device.gatt.connect())
    .then(gatt => assert_promise_rejects_with_message(
      gatt.getPrimaryService('battery_service'),
      new DOMException(
        'No Services matching UUID ' + battery_service.uuid + ' found in ' +
        'Device.', 'NotFoundError')));
}, 'Request for absent service. Must reject with NotFoundError even when the ' +
   'services are not immediately discovered');

</script>
